- made new comps check for steam networks as well - changed steam vent to use damage dealer comp and steamsprayer comp - changed steamsprayer to checck for off devices

+++ Source/CompProperties_SteamSprayer.cs
@@ -10,7 +10,7 @@
 
         public int MaxSprayDuration;
 
-        public int SmokeAmount;
+        public float SmokeAmount;
 
         public float PressureCutoff;
